+++ src/components/Chat/ChatsCardData.js
@@ -1,0 +1,236 @@
+import React from 'react';
+import { Scrollbars } from "react-custom-scrollbars-2";
+import { Card } from "react-bootstrap";
+
+const ChatsCardDtata = () => {
+  return (
+    <div>
+       <div className="row mt-2 d-flex justify-content-center">
+            <div className="table-chat">
+              <Scrollbars>
+                <div>
+                  <div className="col-12  mt-2">
+                    <Card border="success">
+                      <Card.Body>
+                        <a
+                          href="#"
+                          className="list-group-item list-group-item-action d-flex gap-3 py-3"
+                          aria-current="true"
+                        >
+                          <div
+                            className="rounded-circle flex-shrink-0 d-flex align-items-center justify-content-center"
+                            style={{
+                              width: "40px",
+                              height: "40px",
+                              background: "#007BFF", // Set your preferred background color
+                              color: "#fff", // Set the text color
+                              fontSize: "16px", // Set the font size
+                            }}
+                          >
+                            A
+                          </div>
+                          <div className="d-flex gap-2 w-100 justify-content-between">
+                            <div>
+                              <span>Question By Atif </span>
+                              <h6 className="mb-0">List group item heading?</h6>
+                              <p className="mb-0 opacity-75">
+                                Some placeholder content in a paragraph.
+                              </p>
+                            </div>
+                            <small className="opacity-50 text-nowrap">
+                              12-3-2024
+                            </small>
+                          </div>
+                        </a>
+                      </Card.Body>
+                    </Card>
+                  </div>
+                  <div className="col-12  mt-2">
+                    <Card border="success">
+                      <Card.Body>
+                        <a
+                          href="#"
+                          className="list-group-item list-group-item-action d-flex gap-3 py-3"
+                          aria-current="true"
+                        >
+                          <div
+                            className="rounded-circle flex-shrink-0 d-flex align-items-center justify-content-center"
+                            style={{
+                              width: "40px",
+                              height: "40px",
+                              background: "#007BFF", // Set your preferred background color
+                              color: "#fff", // Set the text color
+                              fontSize: "16px", // Set the font size
+                            }}
+                          >
+                            H
+                          </div>
+                          <div className="d-flex gap-2 w-100 justify-content-between">
+                            <div>
+                              <span>Question By Hasnan </span>
+                              <h6 className="mb-0">List group item heading?</h6>
+                              <p className="mb-0 opacity-75">
+                                Some placeholder content in a paragraph.
+                              </p>
+                            </div>
+                            <small className="opacity-50 text-nowrap">
+                              12-3-2024
+                            </small>
+                          </div>
+                        </a>
+                      </Card.Body>
+                    </Card>
+                  </div>
+                  <div className="col-12  mt-2">
+                    <Card border="success">
+                      <Card.Body>
+                        <a
+                          href="#"
+                          className="list-group-item list-group-item-action d-flex gap-3 py-3"
+                          aria-current="true"
+                        >
+                          <div
+                            className="rounded-circle flex-shrink-0 d-flex align-items-center justify-content-center"
+                            style={{
+                              width: "40px",
+                              height: "40px",
+                              background: "#007BFF", // Set your preferred background color
+                              color: "#fff", // Set the text color
+                              fontSize: "16px", // Set the font size
+                            }}
+                          >
+                            A
+                          </div>
+                          <div className="d-flex gap-2 w-100 justify-content-between">
+                            <div>
+                              <span>Question By Ali </span>
+                              <h6 className="mb-0">List group item heading?</h6>
+                              <p className="mb-0 opacity-75">
+                                Some placeholder content in a paragraph.
+                              </p>
+                            </div>
+                            <small className="opacity-50 text-nowrap">
+                              12-3-2024
+                            </small>
+                          </div>
+                        </a>
+                      </Card.Body>
+                    </Card>
+                  </div>
+                  <div className="col-12  mt-2">
+                    <Card border="success">
+                      <Card.Body>
+                        <a
+                          href="#"
+                          className="list-group-item list-group-item-action d-flex gap-3 py-3"
+                          aria-current="true"
+                        >
+                          <div
+                            className="rounded-circle flex-shrink-0 d-flex align-items-center justify-content-center"
+                            style={{
+                              width: "40px",
+                              height: "40px",
+                              background: "#007BFF", // Set your preferred background color
+                              color: "#fff", // Set the text color
+                              fontSize: "16px", // Set the font size
+                            }}
+                          >
+                            W
+                          </div>
+                          <div className="d-flex gap-2 w-100 justify-content-between">
+                            <div>
+                              <span>Question By Wali </span>
+                              <h6 className="mb-0">List group item heading?</h6>
+                              <p className="mb-0 opacity-75">
+                                Some placeholder content in a paragraph.
+                              </p>
+                            </div>
+                            <small className="opacity-50 text-nowrap">
+                              12-3-2024
+                            </small>
+                          </div>
+                        </a>
+                      </Card.Body>
+                    </Card>
+                  </div>
+                  <div className="col-12  mt-2">
+                    <Card border="success">
+                      <Card.Body>
+                        <a
+                          href="#"
+                          className="list-group-item list-group-item-action d-flex gap-3 py-3"
+                          aria-current="true"
+                        >
+                          <div
+                            className="rounded-circle flex-shrink-0 d-flex align-items-center justify-content-center"
+                            style={{
+                              width: "40px",
+                              height: "40px",
+                              background: "#007BFF", // Set your preferred background color
+                              color: "#fff", // Set the text color
+                              fontSize: "16px", // Set the font size
+                            }}
+                          >
+                            B
+                          </div>
+                          <div className="d-flex gap-2 w-100 justify-content-between">
+                            <div>
+                              <span>Question By Boss </span>
+                              <h6 className="mb-0">List group item heading?</h6>
+                              <p className="mb-0 opacity-75">
+                                Some placeholder content in a paragraph.
+                              </p>
+                            </div>
+                            <small className="opacity-50 text-nowrap">
+                              12-3-2024
+                            </small>
+                          </div>
+                        </a>
+                      </Card.Body>
+                    </Card>
+                  </div>
+                  <div className="col-12  mt-2">
+                    <Card border="success">
+                      <Card.Body>
+                        <a
+                          href="#"
+                          className="list-group-item list-group-item-action d-flex gap-3 py-3"
+                          aria-current="true"
+                        >
+                          <div
+                            className="rounded-circle flex-shrink-0 d-flex align-items-center justify-content-center"
+                            style={{
+                              width: "40px",
+                              height: "40px",
+                              background: "#007BFF", // Set your preferred background color
+                              color: "#fff", // Set the text color
+                              fontSize: "16px", // Set the font size
+                            }}
+                          >
+                            J
+                          </div>
+                          <div className="d-flex gap-2 w-100 justify-content-between">
+                            <div>
+                              <span>Question By John </span>
+                              <h6 className="mb-0">List group item heading?</h6>
+                              <p className="mb-0 opacity-75">
+                                Some placeholder content in a paragraph.
+                              </p>
+                            </div>
+                            <small className="opacity-50 text-nowrap">
+                              12-3-2024
+                            </small>
+                          </div>
+                        </a>
+                      </Card.Body>
+                    </Card>
+                  </div>
+                </div>
+              </Scrollbars>
+            </div>
+          </div>
+    </div>
+  )
+}
+
+export default ChatsCardDtata
